Replace LoadLibrary() with LoadLibraryEx()
* also enable LOAD_LIBRARY_SEARCH_SYSTEM32 flag to ensure that
ADVAPI32.DLL gets loaded from the system32 directory on Windows
versions that support this flag.

From https://msdn.microsoft.com/library/windows/desktop/ms684179:
"If this value is used, %windows%\system32 is searched for the DLL and its dependencies. Directories in the standard search path are not searched. This value cannot be combined with LOAD_WITH_ALTERED_SEARCH_PATH.
Windows 7, Windows Server 2008 R2, Windows Vista and Windows Server 2008:  This value requires KB2533623 to be installed.
Windows Server 2003 and Windows XP:  This value is not supported."