Make test_nsalloc_xmlns() robust vs changes in allocation patterns.
Most of the allocation tests use some complicated logic to ensure
that cached memory allocations do not cause the test to miss out
some possible allocation failure paths.  A more robust way to do
that is to free and recreate the parser, avoiding caching
altogether.  This should be done with the test case section
teardown and setup routines, so that the right sort of parser is
created.  This is the first in a series of commits to switch over
to this test structure