Remove speculative test from byte_character_handler().
byte_character_handler() used to test that the string pointer `s`
passed to it was at the correct offset ino the input buffer.  This
is not guaranteed to be the case, and definitely won't be the case
if the input encoding has a different minimum bytes-per-character
to the internal encoding.