introduced single-makefile and partial support for out-of-tree build
Only one "Makefile.am" file is enough for this project, which has a
small number of source files.  The single-makefile infrastructure is
a bit faster and makes it easier to perform parallel builds.

Partial support for out-of-tree builds is also present: everything
that is run using the GNU Autotools should work; the special rule
"make run-xmltest" needs additional work.