Fix -Wimplicit-fallthrough for Clang.
The existing /* fall through */ comments sufficed to annotate explicit
fallthrough for gcc, but not clang.  Convert to a macro that expands to
an annotation in supported compilers.

Further complicating this is that one of the #includes of xmltok_impl.c
makes the fallthrough annotations unreachable at compile time, so they
must be #defined away to avoid yet another warning (for an unreachable
fallthrough annotation) in that case.

Bug: chromium:995993