Fix broken CMake support under CYGWIN.
I tried to build the latest stable of OpenColorIO under CYGWIN, but I got an error because the support for CMake into Expat is broken.
I'm getting this message:

CMake Error at /usr/lib/cmake/expat-2.4.1/expat.cmake:74 (message):
  The imported target "expat::expat" references the file

     "/usr/lib/libexpat.so.1.8.1"

  but this file does not exist.

This happens because CYGWIN uses the traditional DLLs of Windows rather than the ".so" files that you can find on linux.
Hopefully, the fixes are very easy:
1) configure.ac must install expat-noconfig__windows.cmake.in rather than expat-noconfig__linux.cmake.in when it runs under CYGWIN.
2) CYGWIN uses "cyg" as library prefix, while MinGW/MSYS uses "lib". So, expat-noconfig__windows.cmake.in must use ${CMAKE_SHARED_LIBRARY_PREFIX} rather than hardcoding the prefix to the DLL.