Make memory allocation simpler and more uniform
By defining a few overflow-safe memory allocation primitives, most of
the integer calculations are concentrated to very few places that are
easy to review during an audit since the various places where memory
allocation happens no longer do the arithmetics themselves.

The allocation primitives assume that the array length, even though it
is stored as a signed int, is never negative.  They are implemented as
inline functions to allow the compiler to optimize the boundary checks,
especially the division by elemsize.

To make all array handling code uniform, the following items have been
restricted from previously UINT_MAX to now INT_MAX.  These restrictions
are believed to be irrelevant in practice.  The restrictions are:
  - the length of the scaffolding,
  - the number of groups,
  - the number of attributes of a single element