lib: Fix detection of asynchronous tags in entities
According to the XML standard, tags must be closed within the same
element in which they are opened. Since the change of the entity
processing method in version 2.7.0, violations of this rule have not
been handled correctly for entities.

This commit adds the required checks to detect any violations and
restores the correct behaviour.