lib: Split `expat_realloc` to allow a new variant taking the previous size
This is part of a series of changes towards removing the need to track
allocation size as inline metadata. The ergonomics of this new function, taking
`SIZE_MAX` as a sentinel to mean “previous size unknown”, are bad. But this is
only intended to be temporary and will be removed when all callers have been
refactored to pass their previous size.