block struct should have a  Flexible Array Member as part of its struct
Since this is a C99 feature, it works better than a one-member block.

Since this makes it clear that blocks can vary in length, a flexible array member works here.
It also makes it clear that this is not undefined behavior.

Also, yes, this does affect the underlying assembly for the better, by 10 instructions total.

In addition, there is no risk, because:

Size of this struct is not used.
It is not Undefined
offsetof gets the same value anyway.